# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Oauth2 server.",
        "title": "Gandalf API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9100",
    "paths": {
...
        "/notifications/emails/reset-password": {
            "post": {
                "description": "Sends reset password email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "Sends reset password email",
                "operationId": "notifications-emails-reset-password",
                "parameters": [
                    {
                        "description": "resend the reset password email",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.UserResendEmail"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            }
        },
        "/notifications/emails/verify": {
            "post": {
                "description": "Sends verification email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "Sends verification email",
                "operationId": "notifications-emails-verification",
                "parameters": [
                    {
                        "description": "sends the verification email",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.UserResendEmail"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "post": {
                "description": "Creates a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "operationId": "user-create",
                "parameters": [
                    {
                        "description": "Creates a new user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.UserCreateData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/serializers.UserSerializer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/{uuid}": {
            "get": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "user:all:read"
                        ]
                    }
                ],
                "description": "get an user by his uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get an user",
                "operationId": "user-read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.UserSerializer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "helpers.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
...
        "serializers.UserSerializer": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/serializers.userDataSerializer"
                },
                "type": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
...
        "serializers.userDataSerializer": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string",
                    "example": "1997-12-21"
                },
                "email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "name": {
                    "type": "string",
                    "example": "John"
                },
                "phone": {
                    "type": "string",
                    "example": "+34666123456"
                },
                "surname": {
                    "type": "string",
                    "example": "Doe"
                },
                "uuid": {
                    "type": "string",
                    "example": "4722679b-5a48-4e85-9084-605e8df610f4"
                }
            }
        },
...
        "validators.UserCreateData": {
            "type": "object",
            "required": [
                "birthday",
                "email",
                "name",
                "password",
                "surname"
            ],
            "properties": {
                "birthday": {
                    "type": "string",
                    "example": "1997-12-21"
                },
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "John"
                },
                "password": {
                    "type": "string",
                    "example": "My@appPassw0rd"
                },
                "phone": {
                    "type": "string",
                    "example": "+34666123456"
                },
                "surname": {
                    "type": "string",
                    "example": "Doe"
                }
            }
        },
        "validators.UserResendEmail": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://localhost:9100/oauth/login",
            "tokenUrl": "https://localhost:9100/oauth/token",
            "scopes": {
                "app:me:read": " Grants access to read self created apps",
                "app:me:write": " Grants access to write self created apps",
                "user:me:authorized-app": " Grants access an app to get information about the user",
                "user:me:change-password": " Grants access to change self password",
                "user:me:delete": " Grants access to delete self user",
                "user:me:read": " Grants access to read self user",
                "user:me:verify": " Grants access to verify created user",
                "user:me:write": " Grants access to write self user"
            }
        }
    },
...
}